Add collision viewer (#145)
The purpose of this is showing the hex values of the collision layer to
better understand how CheckCollision works. The display has some
compromises due to using FntPrint(). I can't scroll the values smoothly,
and we also lose a few rows/columns due to their FntOpen settings and
some pre-existing prints. There's probably a better solution for getting
text on screen.