DRA ApplyJosephsCloakPalette matching (#368)
Simple and effective. Not sure why we never decompiled this before.

I renamed `g_josephsCloakLiningDark555` into just
`g_JosephsCloak.liningDark` and I believe it is a bit simpler for the
eyes to decode.

I also changed the type `g_Settings.cloak*` to be a single array of `6`
rather than two of `3` given how I matched it in #367.

I also refactored `DarkenCloakColor` as it was a fake match.

Finally I synchronised the symbols between the US and the HD versions.