TT_000 func_8017110C matching (#372)
Took me a day to compile this nasty one. I got a match only by using the
ternary operator `x ? y : z`, otherwise it wouldn't match. I was not
able to get rid of the temporary variables `selfX` and `entityX`. If you
want to play with the scratch you will find it here:
https://decomp.me/scratch/JRqtW