DRA HD DebugUpdate matching and more (#378)
INSANE discoveries today. The HD build found in the game Castlevania:
Dracula X Chronicles is a debug PlayStation 1 build of Symphony of the
Night. It comes with the text viewer enabled by default, controlling the
menu via the second pad, it has a palette viewer, tile viewer, texture
viewer, it can modify the palette at runtime, it can make screenshots,
create video recording and freeze the game to advance frame by frame
with the gamepad and more!

I also took the opportunity to merge the file `42398.c` between US and
HD, revealing some interesting minor differences between the two builds.
Cross-referencing the addresses took me a while. Putting the `#ifdef`s
was more fun.

We now finally have an idea of what some `nullsub` or global variables
do. There is yet a lot to decompile from the HD build, but I assume this
will be the biggest deal we will get out of it.

Please note the `#ifdef rodata` at the end. By cross-referencing
different builds we now have more visibility on when a file split took
place!