Match most of DRA HD 'cd' (#380)
Made more symbol cross-referencing between US and HD. I was able to
match a few new functions thanks to that.

Interestingly in `func_801080DC` I had to make a temporary variable to
make it match on HD. Surprisingly US also matches with it. Pretty sure
there is some fakery happening in there. This is the scratch if you want
to play with it: https://decomp.me/scratch/NNY1O

I was not able to match `UpdateCd` for the time being. The function is
massive. I will do it in a separate PR.