Split saturn libs, add function (#385)
This splits the saturn libs into individual files and adds one function
to test the workflow. The splitter can't split multiple files at the
moment so the linker file is made manually. I plan to fix this
eventually. I updated the splitter to not overwrite linker files that
already exist. I'm open to suggestions on the directory structure but
this seemed like the easiest way given the dosemu copying stuff.