Move Primitive handling functions away from POLY_GT4 (#404)
Per discussion in Discord, renamed func_80107360 to SetTexturedPrimRect.

This raised the issue that both SetPolyRect and this function should use
Primitive as their argument, and not POLY_GT4. Therefore, these are
changed. Additionally, everywhere in the codebase that calls these
functions now uses Primitive for their local variables which get passed
to these functions.

Finally, we update every case of setting blendMode to 8, to use the
BLEND_VISIBLE flag. Hopefully in the future we can identify the rest of
the available blendModes.

---------

Co-authored-by: sozud <122322823+sozud@users.noreply.github.com>