Sound effect extraction WIP (#398)
It seems like there's some interest in figuring out more about the sound
effects lately so I tried making a tool to extract them. They're in a
standard Sony VAB format, split into the header (.VH) and a bundle of
VAG files (.VB). VAG files are ADPCM encoded. Eventually the idea is
being able to round-trip these files but this just extracts to WAV at
the moment. I may be wrong but some of the sound effects appear to be
unused.

Usage:
`sh test.sh`
Look at output in `./out`

The playback speed is hardcoded since I haven't figured out yet if the
pitch is part of these files or somewhere else. It can be changed by
altering `samp_freq`.

---------

Co-authored-by: sozud <sozud@users.noreply.github.com>