Decompile DRA func_800EA5E4 (#421)
I will be completely honest: I have no idea what this function does or
how it works. Something to do with CLUTs.

The function returns 1, 0, or -1, but all the current decompiled
functions that call it do not use the return value, so the function was
previously marked with a return type of void.

I was not able to get a match with the argument being a s32 as it was
before.

This function relies on an array of structs which starts at D_8006C3C4.
Previously, a neighboring function, func_800EA538, accessed what it
believed to be an array of structs at D_8006C3CC. However, it only
accessed a single member of these structs. It was assumed that that
member was the first (unk0), but I have now determined that instead,
that function accesses unk8 of this same struct.

I have therefore renamed the unkstruct to reflect being at C3C4 instead
of C3CC.

There is one very powerful section of code here:
```
    if (temp_v0 & 0x8000) {
        clut = g_api.o.cluts[arg0];
    } else {
        clut = D_800A3BB8[arg0];
    }
```
This indicates that both D_800A3BB8 and g_api.o.cluts hold the same
type. Therefore, we update g_api.o.cluts to hold the new UnkStructClut
pointer, instead of a generic s32 pointer.

UnkStructClut is a new struct I've made. I do not know anything about it
and the only indication of its purpose is that it is set by
g_api.o.cluts. The members in it are just copied based on accesses and
there could be more members which come after the ones I have found.

Unkstruct_8006C3C4 has been modified to reflect the fact that it appears
to have an array that is 0x30 members in size, initialized in this
function. I do not know why it had so many members previously,
especially since the only other function that uses this struct is the
aforementioned one that only touches unk8.

Overall, this function is still in the early days of being understood.
But at least it matches.