DRA+RIC CanTeleportToOtherCastle matching (#449)
Not a very special function per se. I struggled at de-duplicating the
four variables at the top of the function as using just one would give
me some minor reg swaps. I thought adding an enum as a returning value
but I think it is a bit over killing for how small the function is. It
is only called by `EntityAlucard`.

Interestingly, the function in DRA is different between HD and US. But
the HD version is identical to the duplicate found in RIC US. What I
suspect it happened is one of the developers while scrolling the various
functions, found the `- 14407` in an odd spot relative to the rest of
the function and moved it to make it consistent. But as some code from
DRA was copy&pasted to RIC, the same developer did not think to apply
the same change there.

EDIT: Note the `FIX_TO_I`. It is a new macro I added for another
function I failed to decompile.