Decompile DRA func_800FDE20 (#461)
This is a cool one! This runs every frame, and checks if you have enough
EXP to level up. If you do, then it increases your level and gives you
the stat bonuses for it. Therefore, I think `CheckAndDoLevelUp` might be
a good name? I've held off on renaming; if reviewers are okay with the
name then I'll update the PR with this.

I don't know what the 3 values being checked at the beginning are. I
might dig into those next.

D_800A2EC0 is a table which tells how much to increase max HP for each
level. Its values are 1, 3, 6, 10, 20, 30, 40, 50, 100, 200. So the
first 10 levelups you get, you'll gain one max HP, then the next 10 give
you 3 each, etc. I think this should be called LevelUpHPIncreaseTable,
but again, holding off on renaming. I'm not sure if there's a way to
include this as an array in the code rather than as an extern.

I've also included updating the repo to the latest maspsx as part of
this PR, let me know if that should be removed.