Decompile DRA func_800FE97C (#463)
Several comments on this one, it's weird.

This appears to handle some of the logic involved when the player takes
damage. It is called by 3 functions. One is EntityAlucard, and of the
other two one is decompiled and one is not. I have decompiled the other
one, and will submit that PR next.

The first argument is a pointer to a structure which holds some of the
results of this function. This is similar to CheckCollision which stores
the results in a Collider struct. Most of the members of this struct are
still unknown.

The logic with the ballroom mask is extremely strange and is worth
further research. There appears to be a bug since it tests whether the
player is wearing two ballroom masks, which should be impossible.

Overall lots of unknowns being used here, but hopefully this function
and others like it will be helpful in identifying those. I hope I
successfully replaced everything that has an enum.

The line which has `if ((g_Status.defenseEquip > 99)` was causing an
issue with SLTI vs SLTIU, so the type of defenseEquip has been changed
to match this.