Decompile DRA func_801238CC (#485)
In an effort to understand the "entityRoomIndex" Entity member, I have
decompiled this small function which makes use of that value. This
appears to be an entity state machine function, which exists at a
location which is offset from the player left or right, depending on
whether that value is zero, which seems very strange.

I believe I've done the math right here: This function appears in the
0x800AD0C4 array at 0x800AD188. That's 0xC4 bytes deep in the list,
which corresponds to being 49 steps along. Therefore I've named this
UnknownEntId49. We should maybe work out a system for naming unknown
entities (if there isn't one already).