Make the analyze_calls index.html properly use multiple lines (#496)
In the interest of improving the diffs shown in "Update reports" commits
(example here:
https://github.com/Xeeynamo/sotn-decomp/commit/50e507f0c0e0cccfbdaaa83be521f1af3884ee1d),
we recently changed the SVG generation code to be more deterministic,
such that any differences are only the result of actual changes to the
source code of the repo.

While perusing this latest diff, I noticed that the index.html is
showing as a "large diff", which is because the generated HTML code does
not contain any newlines. To be clear, HTML has its own tags which
control the lines in the final rendered page, but the actual text in the
HTML file does not care about whether the code is on multiple lines or
is all in one line.

For the sake of generating diffs, it is heavily preferred for the HTML
to run as multiple lines. With this change, the raw HTML file will be
more intelligible and diffs should be able to properly show individual
changes, rather than struggling to deal with changes that are
technically all within a single-line file.