WRP BlitChar (#528)
![image](https://github.com/Xeeynamo/sotn-decomp/assets/6128729/0c2d011b-c66d-4629-bcae-782124f32f6c)

Used by `EntityRelicOrb` and `EntityUnkId0E` to pre-render the text to
show as a texture.

I am pretty sure the function is bugged. When I use it in the debug
module it runs fine. But when I copy&paste the matching decompiled
function and use it, the game crashes here:
```c
        for (i = 0; i < FontHeight; i++) {
            ptr = chPix;
            ptr += i * FontStride;
            for (j = 0; j < 5; j++) {
                ptr[0] = ptr[1];
                ptr++;
            }
            *ptr = 0;
        }
```
The debug module is using a much more modern compiler, so it is entirely
possible that there is an undefined behaviour in the current logic. I am
poorly convinced about the `if (ch == MINSCODE)` and skeptical of the
`if (i != FontHeight)` condition. The crash happens when a non-wide
character is used (e.g. `l`, `i`, `e`, `d`). It matches though, so there
you go.

The font the game uses is found in the game BIOS and it is retrieved by
`Krom2RawAdd`. According to
[Gemini](https://www.romhacking.net/forum/index.php?topic=17691.20) the
font width is resized in SOTN. This is done by `g_api.func_80106A28(ch,
1)`. Although good luck at decompiling it:
https://decomp.me/scratch/Lulhn

Co-authored-by: @Mc-muffin for figuring out the `while(true)` loop
Co-authored-by: @sonicdcer for figuring out the regallocs