Turn Entity.unk6D into an array (#532)
I have been working on the TestCollisions function. Scratch for
reference is here. https://decomp.me/scratch/gAkz5

We have a line (currently line 75, but that may move) that does `if
(iterEnt1->unk6D[i] != 0) {`, treating an entity's unk6D items as an
array. The codebase currently does not use the unk6E, unk70, or unk74
members, so that's promising. The fact that the for loop goes up to 11
indicates that the size of this array should be 11, which agrees with
the other members that were previously in the Entity struct.

Other functions have referenced unk6D, those accesses need to be changed
to unk6D[0] now that unk6D is an array.

I am hoping to have TestCollisions matching soon. This change will be
necessary to merge that, so I wanted to get this in as its own thing to
reduce how much impact TestCollisions itself will have on the project as
a whole.