Fix function finder svg filenames (#546)
When `analyze_calls` creates an SVG file, the name of that file consists
of the overlay name, then a dot, then the function name. This is
intended to ensure that we get a unique name for the file, given that
the function name alone will often lead to a duplicate. The function
finder was using an incorrect filename-finding logic.

This works on my machine, hopefully it will work in CI too.
Unfortunately this is the type of PR which we won't know if it works
until fully merging. The logic is fairly straightforward though, so
hopefully this performs as intended.