Decompile DRA func_8012D28C (#551)
Mostly just another random DRA function.

This one is a little weird in that it takes an argument which only
serves to make it exit early. When that happens, it also directly sets
the `g_Player.padTapped` variable, which seems very odd to me.

Hopefully something about this function and how it deals with the
g_Player.D_80072CF0 array might help figure out what that array actually
controls. Seems really odd here.