Ci pipeline speedup with cache (#559)
As suggested by @delta-473, GitHub recently shipped their [caching
mechanic for the
CI](https://github.com/actions/cache/blob/main/caching-strategies.md). I
can think of a few ways it can get used in our advantage:

* Cloning the "dependencies": it should save cloning about 4GB of data
to every commit on `master` and about 1.5GB per PR.
* Using this cache instead of artifacts to export the `.map` files to
calculate the decomp progress
* Same as above but to the reporting step.

This PR just addresses the first point as it is the heaviest burden of
the CI so far and it is what I think a quick win.