Generate markdown directory to function graphs (#564)
When we generate all the function graphs in analyze_calls, we generate
an HTML file which contains links to all of them. Example here:
https://github.com/Xeeynamo/sotn-decomp/blob/gh-duplicates/function_calls/index.html
Unbeknownst to me, Github does not allow you to view it as a rendered
HTML page. You can only view it as raw text. There are workarounds like
https://htmlpreview.github.io/?https://github.com/Xeeynamo/sotn-decomp/blob/gh-duplicates/function_calls/index.html
but that's a bit clumsy and you can't get to it directly from the repo.

On the other hand, we have Markdown pages like
https://github.com/Xeeynamo/sotn-decomp/blob/gh-duplicates/functions.md
which render nicely and are usable from Github directly.

This attempts to generate a Markdown page which should be effectively
indistinguishable from the formatting of the existing HTML. Once it's
merged, we should be able to look at the output markdown file, and if we
are happy with it, I will remove the HTML generation, and from then on
this will be Markdown-only.