Decompile DRA func_8010A234 (#568)
A nice DECOMP_ME_WIP to resolve.

The behavior with D_8017A000 is really weird. I copied the handling of
this from func_80109594 which is how I got the match to work.

It seems like D_8017A000 is sort of an API to the currently-held weapon.
It reminds me of the gameApi overlay.

Anyway, it matches, so that's nice.