Change all instances of facing to facingLeft (#574)
A recent discussion on Discord determined that we should change the
"facing" variable to "facingLeft", so that it will be clear in all cases
what is being tested. For example, in lines like `if (PLAYER.facing == 0
&& (g_Player.padPressed & PAD_LEFT)`, the fact that it says facingLeft
makes it clear that the player is pressing in the opposite direction
they are facing (because facingLeft is 0, or false, and they are
pressing left). This makes it obvious that this is some kind of
slow-down or turn-around routine, rather than speeding up or walking
forward.

This was done with a direct find-and-replace of all instances of
`.facing` and `->facing` for `.facingLeft` and `->facingLeft`, followed
by changing the struct members in their .h files. I don't think anything
else got caught in the crossfire here.