SEL StreamNext (#575)
Adapted from `PSX/SAMPLE/CD/MOVIE/TUTO0.C`. I took the opportunity to
rename a few functions according to the sample code too.

All the functions there seems to be self-contained. So I decided to
rename the source file as `stream.c`. Those are the functions
responsible for animating the KONAMI movie logo