Decompile DRA func_8010D800 (#577)
Notes:

1) We are pulling in more .data. Because this is just consecutive after
the previous data, we just bump out the next segment in the splat.

2) Now we're using the unk1 member of the entity1 extension. Cool!

3) We turn out to be using parts of a previously-defined array.
Generally, the code in this function feels a little clumsy. I don't like
temp_t0, temp_t1, and temp_t2, but they're needed. Because I don't know
what g_Entities[1] is, or what it's holding in its unk1 and unk2 values,
I don't know better names for these variables. Same goes for
`current_thing` which is a terrible variable name, but for now, it
serves the purpose. In the future when we can work out more of how this
works, we can give these useful names.