Decompile DRA func_800FA3C4 (#593)
This is a really weird one, and took a long time to work out. As a
bonus, my original decompilation worked on decomp.me with the
gcc-2.6.3-psx and failed with the SOTN preset, which created more
mysteries.

This works a lot with the MenuData struct. Strangely, all of the
accesses are to direct addresses in the struct (rather than storing a
pointer with offsets), with the exception of the if-else block near the
end.

The function, right at the beginning, loads a pointer to one element of
MenuData, but then accesses an element at a 16-byte offset from that
pointer. It seems possible that either the MenuData struct contains an
array, or even might contain another struct (so the overall struct may
be 3 menus, and then a struct we could call MenuDataSub or something).

In any case, after lots of fighting, this matches, and hopefully can
contribute to more research in making sense of MenuData.