DRA func_8012E9C0 and structures (#606)
This is a function which accesses data in an area where we had a partial
struct set up. The behavior of this function helps identify what that
struct should look like.

We have a loop which reads s16 values from a location, but the address
that increments each time we read ends up incrementing by 4, so we skip
a value each time. I decided the best way to describe this behavior is
to create a struct which is a pair of s16's, and then we iterate through
an array of those structs.

The array of pairs is at 0x800ACED0, which falls in the middle of an
existing UnkStruct. Unkstruct_800ACEC6 was created a long long time ago,
and used in func_8010BF64. Working out the details of this struct has
allowed some cleanup on that function, including a `g_Player_unk0C` to
the proper `g_Player.unk0C`.

Ultimately, the struct changes are more important in this PR than the
function decompilation itself. This has some ripple effects through the
rest of the repo.