Decompile DRA func_801177A0 (#609)
More work in gathering data on the variables around D_800ACE##.

Not much progress on what any of those actually do, but this function
seems to match the structures that I established in my previous PR.
Hopefully with enough matching functions that use this data, we can work
out what it's meant to do.

`else_cycles` isn't a great variable name, but... it counts how many
cycles take the 'else' branch. Once we know what values we're comparing
this could probably be renamed.