Decompile DRA DrawJosephsCloakMenu (#611)
This is the pop-up box menu that shows up when you go into settings to
change the color of Joseph's Cloak. I decided to name the function in
this way, which I think is reasonable.

It also uses a unique MenuContext, so I named that as well.

We could probably rethink the way the strings are handled (specifically
c_strExterior and D_800A2D7C) but for now this will match.

Handling of US/HD is a little weird, but nothing too crazy, at least I
was able to get them all to work in one block.