SEL func_801AE6D0 (#621)
Last function of today. I spent much more than I wanted to get a match.
The evil `level = (&g_SaveSummary[port])->level[slot];` was the trick to
get a match all along. On the surface it makes no sense. But it must be
connected to a pattern we have yet to discover. Maybe the developers
were accessing the arrays as `*(array + index)` instead of
`array[index]` (I've seen that being the case in other functions), maybe
they used some strange macros or maybe they used `inline` functions. If
you want to have a play, this is the scratch:
https://decomp.me/scratch/ZePxq

I took the opportunity to give some good names to the dependent
function's arguments.

I am well aware I could have renamed `D_801BC3EC` and `D_801BAF08`. But
as I am in the middle of decompiling
[func_801ADF94](https://decomp.me/scratch/A177U) I prefer to hold back.