Rename transformation exiting functions (#629)
`HandleTransformationMP` is called in EntityAlucard, and by 3 other
functions. One of those 3 was MistFormFinished, and the other two did
similar logic for the bat and wolf forms. Therefore, this PR renames the
other two to the same for bat and wolf.

Happy to change all 3 of these if we don't like the name, but whatever
we do, all 3 should have the same naming structure, and this
accomplishes that.