Decompile DRA func_8010BFFC (#633)
This is an extremely important function and will be very educational.
There are a lot of interesting things.

First, the opening statement has a bunch of conditions that will cause
an early return. They are different between US and HD, which is
interesting. Luckily, the nicest way to write the function (each OR
option on its own line) also makes it very easy to `if defined()` them.

Second, there are two places where I had to use (*(&PLAYER)).posY.i.hi.
I do not know why I can't just use PLAYER.posY.i.hi, but when I try, the
addressing of the variable gets all messed up. Feel free to try in this
scratch: https://decomp.me/scratch/Rjvwe

Third, this function has two big IF statements, where the body of both
statements is mostly the same. The body is a `while` loop. I couldn't
get the flow to work without it being a `while(true)` with `continue`.
This is a bit sad, but maybe this is how it was really written.

Fourth, and most important: This is passing `g_Player.D_80072BD0[i][0]`
as a Collider to `CheckCollision`!!! I think this means that this was an
array of Colliders all along! In fact, a Collider is 0x24 bytes (36
bytes), and currently we have `s32 D_80072BD0[8][9]` - and s32[9] is
also 36 bytes! I think this should be `Collider colliders[8]`. This is
very very cool!

I'm going to leave the structure alone for this PR, because clearing up
all instances of D_80072BD0 will be messy. But longer-term, this should
help a whole lot in understanding the behavior of these members of
`g_Player`! Very very exciting!