Decompile w_029 func_ptr_80170028 (#646)
We're currently trying to resolve some issues with compiling/assembling
w_029. This is a function in that weapon which I was able to
successfully decompile.

Frustratingly, we named unk7C to lifetime since we saw a different
weapon using it this way, but now we see this using unk80 the same way
(with a lifetime of 10, decrementing on each run). Not sure what the
best way to handle that is, but hopefully we can work out more patterns
as we work out the various weapons.