Change `li` macro; decompile w_029 func_ptr_80170004 (#648)
This was a huge pain to get working, but we got it. With a lot of help
from @Xeeynamo and @mkst , we figured out that there was a weird edge
case with the li macro and loading large values. That is now fixed.

The value we need to load is 0xCCCCCCCD. There were two issues
associated with the macro's parsing of this value, relating to the fact
that the assembler treats all values as signed, which causes issues both
in the if statements, and in the behavior of the %hi. Of the two changes
in the macro, only the lower one is strictly needed for this case, but I
added the upper case too, since i expect we will run into it at one
point or another.

Anyway, this works! Yay!