Decompile w_029 EntityWeaponShieldSpell (#649)
Now that we fixed the `li` instruction, the other functions in this
weapon file can come spilling out :)

There was a jump table that I changed to being rodata. Luckily that went
smoothly.

Interestingly, this does the same %5 as the other function, but there
are two spots where it uses g_Timer (the one that always runs) and one
other where it uses g_GameTimer (the one that pauses when the game
pauses). It's only palettes so it doesn't really matter, but it's neat.

One problem: This stores two bytes to self->7C and self->7D. But we
already had a different weapon that used 7C as a s16. So we may end up
needing to make different `ext` entries for the different weapons. For
now, I'm leaving it generic.