Decompile w_029 func_ptr_80170024 (#651)
This one was big and complicated. It rounds out w_029; there is no
longer any INCLUDE_ASM in this, so that's cool.

I believe this function controls the shield rod spell with the fire
shield, but I can't guarantee that. This required creating a new struct,
which I called FireShieldDragon. I put it in the .c file since I don't
expect anything else to use that struct so it doesn't need to be in a .h
file.

Overall this function could probably use some study and cleanup, but for
now it works and it's great to have the whole function decompiled.
Nobody will ever have to deal with the `-O1` of this function ever again
:)