Decompile DRA func_80118970 (#661)
This one took a while but it works now so that's nice.

It seems to search for entities matching a certain set of criteria. 

The D_80138038 variable seems to only be used in this function. I don't
know what it's used for.

I also don't know what the point of the behavior with temp_a0 and var_v1
is for. We have a lot of dividing and multiplying by 128 (which is also
the size of the array we're looking through).

For some reason without the LOH() I was loading posX.i.hi and posY.i.hi
with `lhu` instead of `lh`. Not sure why, and not sure why I only need
to do that on one of each of those, but either way it works so that's
good.

Overall lots of mysteries with this one but hopefully having it
decompiled will be the first step toward answering those mysteries.