DRA func_800F84CC (#685)
This function seems to be related on how to display certain menu
graphics. The only affected visual element I saw is the preview icon at
the bottom right in the Equip sub-menu. I was not able to identify how
this function fits in the grand scheme, therefore I did not find any
suitable name.