Decompile DRA func_800F99B8 (#691)
I really don't know what this one does, which is why all the variable
names are placeholders. But this matches, so hopefully others who are
more familiar with shift-jis and similar structures can determine what
this is doing.

Main reason for doing this one was to help with the effort to finish
5298C.c.

Next I will attempt doing func_80106A28, which is called by this
function and is very complicated.