Decompile DRA func_8011BBE0 (#696)
Not sure what this one is. The entity is ID 24. This comes right after
EntityHolyWaterFlame in the entity updater functions list. I tried
setting a breakpoint and playing around with subweapons for a bit but
couldn't get the breakpoint to trigger.

I'll keep trying to play with this to figure out what it is, but for now
we can at least check that my code is good.