Decompile some duplicate functions (#697)
My last PR involved decompiling an unknown entity. I was hoping to find
a call to CreateEntityFromEntity which would match, but none has been
decompiled yet.

I noticed that there is a function which calls CreateEntityFromEntity
which is decompiled in all overlays except CEN and RWRP. I figured we
might as well finish it out and have it done in all of them.

The RWRP version calls AllocEntity, which also wasn't decompiled in
RWRP, so I copied that in as well.

Not at all an important PR, and if we're not doing PR's that are only
duplicates then feel free to close this without merging.