Rename PLAYER_facing symbol to PLAYER_facingLeft (#779)
A while ago, we renamed Entity.facing to Entity.facingLeft so that the
binary state would be obvious at a glance.

The symbol for PLAYER_facing was never renamed. When I create scratches,
one of the first things I do is a find-and-replace for `PLAYER_` to
become `PLAYER.`, and having the symbol match the new entity member name
will make that much more convenient. Ultimately, this PR is tiny and
just makes things a bit more consistent.