Decompile DRA EntityGiantSpinningCross (#794)
This was a weird one. This uses the Geometry Transformation Engine,
which I don't think any previous function has done. It was confusing to
get it working but it seems to work.

I think I put the macros in the right place, but let me know if not.

Overall this is the first PR of this type, so please do make sure
everything looks good so we don't pollute the repo. If this goes in, I
might try to find more of the gte functions so we can crack those all
away. Making this work was difficult and required a lot of SDK research,
so I'm hoping I can do more functions before the knowledge disappears
from my brain :)

---------

Co-authored-by: sozud <122322823+sozud@users.noreply.github.com>