Allow main roundtrip across different states (#804)
The most prominent change is the new macro which replaces the `DEBUGF`
for each stub, but that is the least interesting part. This allows to go
through the different steps `Game_Init`, `Game_NowLoading` and once it
finally lands into `Game_Play` it calls `func_800F298C`, which is our
new `main` function for testing stuff.