Decompile RIC func_8015EE28 (#843)
This one is messy but it works. Another of those big collider functions.

Had to split g_Player.colliders into two arrays, so now we actually have
3 arrays of colliders going on. Might have more splitting to come.

We have `g_Player.colliders2[i].effects >> 4` which used the wrong
`sra`/`srl` instruction, so this indicates that it needs to be u32, not
s32.