Decompile RIC func_8016F198 (#928)
This one is pretty weird, there are a whole bunch of temp variables that
I couldn't figure out how to solve.

There are 8 blocks which set a variable, capped between 0 and an upper
limit (0xF0 or 0xFF). For the upper limit, I have created a `MIN()`
macro. I tried to make a `CLIP()` which would handle both limits in one
macro, but couldn't create one that would match.

Scratch here if anyone can improve things:
https://decomp.me/scratch/fY8js