Decompile DRA func_80136010 (#943)
One of the few remaining DRA functions.

Some of the code here seems fake, but it matches so I'm going with it. I
tried to change all the pointers to use the loop indexing but it didn't
work. Might have been written by a coder who was more comfortable
working with pointers than with loops.