Fix load of ST/SEL/F_SEL.BIN (#972)
The file is 192KB long. Yet on PSX it seems the game still wants to load
256KB of it. Obviously that only works there where the memory is
unprotected.

There are two other crashes happening at `func_801B1FD8` and
`func_801B2108`. The first one is triggered by `func_801ACF7C` and it
seems to be related to the termination character. The second I have no
idea.

To skip `SEL` and boot straight to WRP, go to `sim_pc.c` and comment the
following lines:
```c
        case STAGE_SEL:
            InitStageSel(&g_api.o);
            break;
```