Null backend (#900)
This implements a null backend without SDL2 that can be run in CI and
also serve to show what the current platform-agnostic API is. There is a
core library, which is DRA and the pc supporting code, and two targets,
the SDL2 target, and the null one. null.c is the API that has to be
implemented to make a new target.

Since main is an infinite loop I added a little code to make it exit
after 60 frames.

I don't know what is wrong with the headers in null.c.