Implement DrawOTag (#1004)
This was one of the main reasons why I did not want to create a custom
SDK. But after reading online [how the ordering table
works](https://psx.arthus.net/sdk/Psy-Q/DOCS/TECHNOTE/ordtbl.pdf), I
just figured it would be a piece of cake to implement it.

Ohbboy I was wrong.

It took me half a day to reach this state. But this works on 32-bit and
64-bit builds without hiccups. Implementing `DrawOTag` was to me the
biggest obstacle. Please enjoy a series of screenshots that gives the
idea of how important this function is:


![image](https://github.com/Xeeynamo/sotn-decomp/assets/6128729/e4e9f081-1342-4441-b2d4-2c73993ada7b)


![image](https://github.com/Xeeynamo/sotn-decomp/assets/6128729/1befcacd-60f4-4492-98b6-b05ddbed228b)


![image](https://github.com/Xeeynamo/sotn-decomp/assets/6128729/59eab278-59da-47b6-b1ee-67e8604e4305)


![image](https://github.com/Xeeynamo/sotn-decomp/assets/6128729/df7e1e43-985b-44b1-8436-aca35c77b7ac)