Decompile RIC func_801682B4 (#1014)
Exact copy (aside from using g_api) of an existing DRA function. It was
not obvious that these would match, and took a lot of digging to work
out that they were the same, but I'm glad I went through that process,
because otherwise decompiling this would be annoying! Definitely will be
important to keep doing that as we go forward. It's amazing how much of
RIC and DRA are similar.