Decompile DRA and RIC EntityAguneaHitEnemy (#1030)
This one was pretty complicated but I'm glad I got it working. Not
completely certain what this does but good enough for now. The two
functions are even more similar than DRA and RIC functions usually are.

For any comments to be made, feel free to only comment on one of the two
instances and I will make the change on both.