Decompile ST0 func_801AF380 (#1045)
This is some sort of helper function, called by the entity func_801AF774
(not yet decompiled).

It's kind of a mess, but unfortunately everything I tried to make it
pretty broke the match in subtle ways.

Because it is only called by one entity, I make a new Ext member for
that entity, since this plays with prims in the Ext section. I will next
proceed toward decompiling that entity, which will likely involve making
a more complete Ext entry.