Render the explored castle map (#1081)
![image](https://github.com/Xeeynamo/sotn-decomp/assets/6128729/fad98e8c-ff42-4944-b54f-7e97f46dd98c)

This. It still misses the texture and I do not know what the green lines
are about. All the stuff within the `#else` and `#endif` are
placeholders that will hopefully be fixed later on the same way.

To get the same result from the screenshot I had to use the
[func_800F1B08](https://decomp.me/scratch/nW7lV) scratch, comment
`g_CastleMap[i] = 0;` from `InitStatsAndGear` and use the following to
read the map from a memory card's save:

```c
if (!FileRead(LoadMemcardSave, "memcard1.mcd")) {
    WARNF("unable to correctly load the save file");
}
```
```c
s32 LoadSaveData(SaveData* save);
bool LoadMemcardSave(FILE* f) {
    const int slot = 0;
    SaveData save;

    fseek(f, 0x2000 + slot * 0x2000, SEEK_SET);
    fread(&save, sizeof(SaveData), 1, f);
    return LoadSaveData(&save) == 0;
}
```