Decompile weapon 49 func_ptr_80170014 (#1123)
My last PR was for weapon 15, this is the same function for weapon 49.

Substantially similar; interesting that this one uses a switch on the
self->step while the other one just had an `if step == 0`.

Otherwise nothing much of interest here. Going to continue to leave
things as externs until each weapon is done.