Decompile weapon 21 (food) func_ptr_80170004 (#1268)
Unfortunately, this required an evil change of the return type, even
though it doesn't return. And since it's one of many weapons that share
the functions, they all had to change.

Otherwise, nice to have the food figured out (this is the last food
function), since food is a somewhat major part of this game.