Defake NP3 func_801CD78C (#1319)
This function is part of a small network of functions that pass around
Entities and Point32s, so it ends up being a little weird. I guess the
first two elements of an Entity are equivalent to a Point32, but still
weird.

Within this function though, there is nothing to indicate that the args
are Entity, and a PSP function seems to allocate the right stack size to
call this with Point32, so I think we should do that, and have the
callers cast their entities to Point32 if needed.