Minor cleanup on wolf function func_8013136C (#1378)
This is a relatively unimportant function, but it messes with some
variables that I am interested in learning more about, so I went ahead
and compared it to the PSP version.

This PR incorporates some lessons I learned from that PSP version. Most
importantly, the rotPivotX and rotPivotY in Entity are signed.

I also tracked down how this entity is created, it's a bit odd since it
doesn't call any entity creation function, and instead does a direct
write into g_Entities to create it.

On exit, this function calls `func_8012C600`, which I will be doing a
similar analysis for next.

There was a leftover rodata padding that doesn't do anything, so I
removed that too.