Add texture viewer tool (#1570)
A tool I made that allows you to identify some entities based on the
primitives that they create. Usage is documented at top of the file.
Allows extraction of images, which has already been useful in a few
places.

Happy to provide any assistance with using this; it's come in handy for
me and hopefully will be beneficial for others too.