Fix RBO3 progress (#1810)
The code path I deleted was pretty much dead with the exception of
`BORBO3` where the bug was triggered. Essentially the tool was looking
for `asm/us/boss/rbo3/nonmatchings/rbo3/nonmatchings`, which had no
content and let the tool assume the overlay completion was 100%.

I also added a guard where if `nonmatchings` is not found but also
`matchings` is not there, the tool will terminate immediately. This
should prevent any future oddity.