Decompile ST0 func_801BC5C0 (#1832)
A very ugly one with a ton of GTE instructions that required fixing M2C
errors to get it to work. Glad to have it done!

Had to do a file split due to EntityCutscenePhotograph in the same file
having rodata that isn't pulled in yet.

This uses a few GTE macros which weren't being used before. While I'm at
it, I reorganized the GTE macros file to match the order of the
functions as used in the SDK's INLINE_C.H file. Doesn't really matter,
but the file has some light organization so I figured we might as well
match up with it.