Decompile ST0 EntityBackgroundVortex (#1838)
Another one down!

In terms of the sequence of functions, this is the last function in the
ROM for ST0. The split between C and BSS was off by one (the NOP after
the RA for this function was being pulled in as part of the 00 bytes in
BSS), so I had to make some small BSS fixes to get a match. All good
now!

I believe this leaves only a single ST0 function remaining.