Data Import tt_002  (#1896)
Importing the remaining data for tt_002
Importing sbss data
Adjusting the data files

The data except for the servant_events data is all unique to the faerie.
I suppose all the data could be put in 1 file, but I had it split up
because ghost and bat shared a significant amount of data.

~~D_us_801724CC[] should actually be an array of Unkstruct_801724CC;
however, that doesn't match because it's on a 2 byte barrier and the
compiler wants to align it to 4 bytes. I tried #pragma pack and
__attributed__((packed)) and a few other things with no avail. So that
and the cast to Unkstruct_801724CC in func_us_80177F84 are just going to
have to be wonky.~~
This issue was figured out by @Xeeynamo 

I didn't really do any refactoring beyond what I needed to in order to
get the data in. I'll work on that next.