Changing PLAYER_STATUS_UNK40000 to PLAYER_STATUS_ZERO_HP (#1911)
Simple change. Figured out that UNK40000 is either "Player has Zero HP"
or "Player is dead", which is largely the same I think. The Faerie "Life
Apple" ability triggers off of this status. Touches a lot of files so I
didn't want to put it in my Faerie refactor.

Also a small renaming in Faerie that I realized I forgot in my refactor