Decompile `no1` `e_blue_axe_knight` (#2017)
Apologies for the size, this ended up being bigger than expected the
more I picked at it.

Decompiles the blue version of axe knight found in Outer Wall, Castle
Keep, Colosseum. This enemy is very similar to the existing green axe
knight header we had, but has some additional attacks and an extra
function.

As I decompiled it, I found it easier to separate this into a separate
entity rather than trying to fit into the current shared header with
conditionals.