Update/document game logic for handling elemental resistances, absorbs, etc (#2020)
Lots of cleanups here with renaming stuff, creating enums, and more.

This was the kind of thing where I kept on having a cycle where I would
improve all the uses of a certain field, then in doing so, I'd see where
it was used, rename the function, notice more fields, and so on. But I
think in the end I tied up the loose ends that I found, and this becomes
a nicely contained PR which should help lots of study.