PSP DRA func_800F087C and an important file split (#2028)
86A0.c is the HUD file I've been doing. This is the next function in
that file. However, it's completely unrelated, is not the next one in
the US version, and has different compiler optimization (-O0 versus
-O4). This heavily indicates that there should be a file split here. So,
this PR decompiles the function, swaps out that decompilation for the US
version, and performs the file split to continue preparing for a future
`hud.c`. Nice to see that we now have the two versions ending the HUD
file at the same place, now just need to get the starts to align the
same way.