PSP for DRA func_801028AC (#2039)
I got stuck on HUD stuff so pivoting to this. This is the first function
after one of our shared cross-platform C files.

I couldn't get one of the if-conditions to work in both systems, so it's
changed to an ifdef for now until someone finds a solution. It's taking
a decrement and either pre-comparing to zero or post-comparing to -1.

I decided to unify D_801379A4 and D_801379A8 into an array since there
was an optimization that was keeping a pointer around for ...A8, so that
indicates it's a member of an array or struct. Further study will likely
adapt this over time.

But for now, yet another function matching on both systems. Tricky one
since it was O4.