Decompile more ST0 cutscene (#2075)
The `Dialogue` struct is 4 bytes shorter on PSP. It's hard to tell which
field is different without decompiling `ST0_EntityCutscene`. But I
re-adjusted the offsets based on what I think the struct is shorter. The
`g_Dialogue.prim[5]` is infact missing on PSP.

Data is imported using `CH(x)` instead of `_S(X)`. This is because
`sotn-str` cannot yet be integrated on the build-chain on PSP. A longer
discussion about it happened here:
https://discord.com/channels/1079389589950705684/1079395311350452255/1309620547696595015

Some functions were far too different to be ported to the shared
headers.