Prepare for next PSP code share (#2159)
This PR aligns PSP and US to prepare for code sharing, as I've done with
the last few files.

The next PSP function is EntitySubwpnThrownDagger. In US, that function
was in the middle of a file, and came right after func_80124A8C. In
order to make sure this is a real file split that exists elsewhere on
PSP, I dug into this function. It turns out that this handles the Z's
that come out when Alucard falls asleep in a chair, so I renamed the
function to match. I got a little carried away chasing that down, but it
means we have better documentation of this stuff now.

I also went through and made PSP symbols for every DRA entity, which
makes it a lot easier to cross-reference C code now for files that
aren't yet matched up with each other.