Update Splat to 0.27.3 (#2165)
A few things to keep up with breaking splat changes in more recent
versions:

* extensions are now derived from `Segment` instead of `N64Segment`. the
latter did not add any additional behavior to `Segment` so only the
reference needed to be updated.
* removed the `include_macro_inc` option from all splat config.
* `macro.inc` was being included in every extracted ASM file, so include
guards have been added to avoid duplicate symbols.

0.27.2 is the latest version which continues to work without additional
changes. 0.26.1-0.27.2 include a bug triggered by HD DRA where a segment
becomes its own sibling. 0.28.0 requires additional changes that may be
trivial, but increase the scope of what needs to be changed. 0.27.3
seemed to be a nice version for a reasonable incremental change.