Clean up `lib` imported data (#2235)
Data seems to be aligned to a 4 byte boundary, so a lot of imported data
has trailing zeros that are actually padding between data. Here I mostly
remove trailing zeros from `u8[]`s passed to `AnimateEntity` since that
function ignores everything after a `-1, 0` or a `0, 0`.

I also adjusted some imported arrays so their size matches what the code
expects their size to be. I had to add a `STATIC_PAD_DATA(4);` but
perhaps that could be a clue to adjust the file boundary.