Ensure CI pipeline runs after Makefile changes, plus add Python 3 on macOS (#2276)
I noticed in #2258 the other GHA pipeline was not executed as it was not
listening for any `Makefile*` changes. ~This hidden a build fail on
macOS, where Python3 seemed to not be installed.~

~The PR aims to fix both issues.~

EDIT: The PR will fix only the first issue.

EDIT 2: [we just
decided](https://discord.com/channels/1079389589950705684/1135205782703570994/1345798893958463528)
to tweak the Makefile set-up