Decompile BO4 Functions - Doors & Alucard Analogues (#2288)
Adapts several functions from `DRA` to their Doppleganger specific
versions. Functions that could be shared were pulled out and shared.

Many of these functions are the same except which entity and global vars
are being pointed to. There is an off-by-one difference in the steps for
Doppleganger vs. Alucard which shows up occasionally. Entity flags are
often different because Doppleganger is not tracked by the camera.

Includes a door function as well. 

I had correctly spelled Doppelgänger for the `DOPPELGANGER` macro, but
Doppleganger is misspelled in the game, so I've corrected the spelling
to be consistent with the game.