Next file of code sharing for DRA PSP (#2289)
Almost every change here is simple, reasonable, and straightforward.
Nice to see a clean diff; I'm really glad we're using this workflow of
pre-aligning the file splits, and then decompiling in the next PR.

Also should resolve the matter of `BTN_WOLF`, I raised the question in
this in #naming in the Discord.