Added sotn_diff.py and renamed permuter (#2287)
I renamed sotn_permuter.py to permuter_loader.py to make it less easy to
get the filename mixed up with the folder name and the
decomp-permuter/import.py symlink.

I made very small adjustments to
https://gist.github.com/sozud/3d3ea081de6906b6f99b9b336e4de507 so that
it supports passing a bin file and map file to use to compare build to
expected. Uses VERSION or passing the version manually, defaulting to
us, to determine where to look for the files.

I've used it for both us and pspeu, so I know it works, but not
extensively so there are probably pitfalls in places.

---------

Co-authored-by: ProjectOblivion <ProjectOblivion>
Co-authored-by: sozud <122322823+sozud@users.noreply.github.com>