Update `st0` PSP prologue cutscene for match (#2308)
I have chosen not to merge these files for now. There was quite a bit of
difference between the special `g_Dialogue` used on PSP and PSX side,
and many differences in the implementations due to PSP's 5 languages
etc.

Where possible I have "ported" the code cleanups back to the PSX
version. Also imported the PSX script for the prologue cutscene roll
which is what this code is responsible for.

Think this leaves just one remaining ST0 function for PSP
(`ST0_EntityCutscene`)