Rename functions related to revealing secret passages (#2350)
There are a handful of functions that are related to drawing pixels on
the map to indicate hidden passageways that the player has revealed
(like the breakable wall and floor in Alchemy Lab). The array at
`D_800A2BC0` lists out the possible points and what castle flags they
pertain to.

There does not appear to be an entity that reacts to the presence or
absence of the Hidden Stairs in Reverse Keep, so toggling this flag
appears to do nothing. Nevertheless, the pairing of bytes in the
`D_800A2BC0` array implies that castleFlags `0x94` and `0x9B` are
related. The same goes for the castleFlags pertaining to the Elevator in
Colosseum.

Also, I apologize for disabling the formatter around D_800A2BC0, but
with it on it made the grouped relationship of the fields much harder to
see.