Decompile several functions of PSP DRA menu (#2403)
This file is truly gigantic, definitely not going to try to do it all in
one PR.

I started with just doing MenuHandle, but then there were a whole bunch
of other functions that I had to get working first before I could get
that one to come in, due to weird stuff with data.

But now they're all matching, and we have a big batch of new symbols to
use which should be helpful all around.

I might continue with this or move onto something else for a bit, these
functions are really annoying with the O4.