RIC PSP: match some subweapon functions (#2407)
This took me a while. It matches around 25% of the whole RIC overlay on
PSP and it is merged with PSX. There are few differences between PSP and
PSX, but matching between the two required a lot of unexpected code
massaging.

The biggest news is that `u16 hitboxOffX;` is now gone from `struct
Entity`.