Rename animFrameIdx to pose and animFrameDuration to poseTimer (#2410)
I found multiple `FntPrint("pl_pose:%02x\n", PLAYER.animFrameIdx);`
scattered around the various overlays, and we decided to strip the `pl_`
prefix as it is probably only intended to be used for `g_Entities[0]`.
Same for `animFrameDuration`, where `ptimer` was found from
`FntPrint("pl_ptimer = %d\n", PLAYER.animFrameDuration)` on PSP. I
assume it is a short for `poseTimer`.

I think `pose` and `ptimer` make perfect sense in this context. In an
entity animation, each frame represents a pose and each pose has a time
to live.

I decided to not rename anything else that uses `pose` and `ptimer` as I
did not want the PR to be a scope-creep and hard to review.