Decompile `no4` func_us_801D96FC (#2429)
PSX: https://decomp.me/scratch/1JmyJ
PSP: https://decomp.me/scratch/JQPXI

I created a new scratchpad macro that *could* replace `SPAD`. Let me
know if it's a good idea or if I should revert to using `SPAD`.

This was the largest non-decompiled NO4 function, and was pretty tough:

- `points` and `sxy` were originally 4 separate pointer variables while
I was decompiling on PSP, but PSX requires them to be arrays. The stack
variable offsets *grow* downwards on PSX, except for variables in
arrays, while they *grow* upwards on PSP even in arrays.
- `&g_Entities[self->ext.et_801D96FC.unkA8[j]]` looked like
`&g_Entities[(self + j)->unkA8]` until I understood there was an array
in the extension struct.
- `func_us_801D9560` / `func_pspeu_09255F98` had to be declared static
in the same file to get a match on PSP
- Figuring out scratchpad variables is always a brainteaser but so
rewarding when it all makes sense.