PSP RIC: pl_blueprints (#2445)
A monster of PR. Matches several PSP functions and merge them with the
PSX overlay. You can see how PSX functions changed by exploring the
individual commits. I did not want to duplicate functions and merge them
separately, therefore I am submitting everything at once.

The name `pl_blueprints.c` comes from the fact this function mostly
creates entities from the blueprints. I am unsure if the name is really
that accurate with the content, but it's the best I could find. I am
surprised there are so many functions in one file, but cross-referencing
PSX and PSP suggests that being the case. I could mark different
functions as `static` that were not previously possible. Not all the
self-contained functions can be `static` as the PSP code will start to
not match. All the weird fake-ry code was removed too.

This decompiles around 24% of the RIC overlay. We're almost there!