PSP RIC: import remaining data (#2480)
With this PR we will have all the `.text`, `.data` and `.rodata`
matching and shared between PSP, PSX US and PSX HD!

The key change was to inline the previously extracted asset
`subweapondefs` as C code. Given the data and C order between PSX and
PSP, it is clear the struct array was meant to be inlined. At the same I
did not want to lose @bismurphy work. I moved the asset
extraction/compilation from a Splat ext into the asset tool, with the
objective of expanding the support for future moddability. This is now
how the exported `subweapons_def.yaml` looks like:

```yaml
PL_W_DAGGER:
  attack: 20
  heartcost: 1
  attackelement:
  - ELEMENT_CUT
  - ELEMENT_HOLY
  chainlimit: 3
  nframesinvincibility: 4
  stunframes: 4
  anim: 0
  blueprint: BP_SUBWPN_DAGGER
  hitboxstate: 2
  hiteffect: 129
  crashid: 21
  unk11: 0
  entityroomindex: 32
```

This gets compiled into the following source code (excluded from git):

```c
{20,1,ELEMENT_CUT|ELEMENT_HOLY|0,3,4,4,0,BP_SUBWPN_DAGGER,2,129,21,0,32},
```

I am leveraging the existing enums to parse the elements and blueprints,
with the goal of keeping the assets synchronized while maintaining a
human-readable YAML file.

I am not yet applying this to `DRA.BIN` as that is still largely
work-in-progress.