Refactor AnimationFrame definitions (#2515)
A change I wanted to do for a long time.

I noticed how the second parameter of `FRAME` was always an even number.
That is because the macro was allocating the two parameters on the wrong
bits. With the discovery of `pose`, I took the opportunity to make a
dense refactor that will help the creation of mods.

I choose from `FRAME(x, y)` to `POSE(z, y, x)` as duration and the old
FRAME were always going in pair. I see no value to keep duration
separate.