Asset tool revamp (#2520)
* All the generated source files will now have a `gen_` prefix in their
file name. This is important when extracting new assets as it gives
freedom to assign them any name instead of having to rely on a pattern
or to keep polluting the `gitignore`. It also makes it easier to
understand at glance which files are generated by the tool
* Write a custom PNG encoder. Unfortunately the one built in Go perform
pre-multiplies the palette colors with the alpha channel, effectively
making it lossy. For example `R: 255, G: 255, B: 255, A: 127` will be
saved as `R: 127, G: 127, B: 127, A: 127`, which is wrong. Writing the
PLTE and tRNS headers manually will allow tools such as Aseprite and
GIMP to preserve the information.
* Two very minor adjustments on versioning and path concatenation