Match NO4 func_us_801c86f4 (aka EntityBreakableCrystalFloor) (#2525)
In room ID 1 of Underground Caverns (which I call Hidden Crystal
Entrance), there is a breakable floor at the bottom made of crystal (or
maybe it's ice?). After striking it three times, the tiles completely
break away, leaving a hole that the player can fall through to enter the
room below.

The function references an array of data that stores the tilemap IDs to
use when updating after each break attempt. This array should look like
the following:
```c
static s16 D_us_80181880[] = {
    // Phase 0
    0x054C, 0x054D, 0x054E,
    0x0180, 0x0551, 0x0552,
    // Phase 1
    0x054C, 0x0740, 0x0748,
    0x0180, 0x0551, 0x0552,
    // Phase 2
    0x054C, 0x074D, 0x074E,
    0x0180, 0x0551, 0x0552,
    // Phase 3
    0x055E, 0x0000, 0x0000,
    0x055F, 0x0000, 0x0000
};
```
I have not tried to add this array as part of the PR, but am identifying
it in case it proves helpful for someone else later.

Special thanks to @JoshSchreuder for helping me figure out the final bit
of the match on this function.