Rename Doppleganger Functions That Overlay DRA (#2534)
Several functions in `bo4` and `rbo5` were deduplicated to names which
overlay with those in `DRA`. At a minimum these operate on
`DOPPLEGANGER` and `g_Dop` instead of `PLAYER` and `g_Player` so are
need to be unique but should not share a name with the equivalent player
functions. This change renames anything with `Player` to `Doppleganger`
to differentiate these.