Decompile PSP MainGame (#2536)
Code-sharing this one would involve more work than usual, because we
will need to re-split PSP and PSX. This is because right after MainGame
there is a GetLang function, but that's static, and there is already one
in the file that MainGame is in. That indicates to me that after
MainGame we should have a new file, to avoid having two GetLang
functions in the same file.