Decompile `dra_psp` play.c (#2539)
I started importing some of the data but `loading.c` depends on data in
`play.c` indicating that they are the same file. Unfortunately there are
multiple `GfxEntry`s between the 2 files that needs to be imported first
and those have type `GFX_BANK_COMPRESSED` which we don't have an asset
importer yet.