Decomp/codeshare PSP NO3 Entrance Weights (#2562)
This is the pair of weights in the upper right next to Cube of Zoe which
opens the shortcut to Marble Gallery. Not too terribly many surprises in
the process of solving this, but some small code details cleared up by
PSP.

Too bad the built in diffing can't identify the old versus new code when
it gets pulled into a .h file.

---------

Co-authored-by: sozud <122322823+sozud@users.noreply.github.com>