PSP NO3 Jewel Sword Rock Puzzle, decomp/backport/codeshare (#2575)
The data for all the rock tiles is handled very weirdly in this
function. Making it work on both versions was very tricky. It's nice
that PSP aligns data to 8 bytes, since that makes it clearer where one
array ends and another begins, due to zero-padding existing that does
not exist on PSX.

Anyway, this works on both versions!