NO3 PSP Trap Door (#2580)
Had an interesting discovery that we had the data splits wrong on US, so
some data was being included in the trap door animation array (past its
0xFF terminator). That data actually belongs to the jewel sword puzzle,
which also meant we removed the ifdef for PSP. PSP had all the jewel
sword stuff right, and US had it wrong.

All better now.

---------

Co-authored-by: sozud <122322823+sozud@users.noreply.github.com>