Extract VH and SEQ files (#2593)
Now these files are extracted in `assets/dra` with the correct
extension.

I moved the asset extraction from splat to the asset tool, so doing a
`bin/sotn-assets config/assets.us.yaml` will be enough to extract the
files without using splat. This will allow us to remove some logic from
`make extract` to make the CI passing.

The various sequence files are imported via `__asm__` to skip the
aligement by 4. GCC 2.6 does not support __attribute__ packed, and this
is the only way I found to get a match.

I took the opportunity to rename a couple of files to better create some
harmony with the naming