Find Segments and More Symbols in `make-config.py` (#2387)
This change updates `make-config` to look at several reference overlays
for finding overlay boundaries and the addresses of their symbols in the
new config being generated.
    
`make-config` uses a set of "reference" overlays to generate
fingerprints for segments as well as functions within those segments. It
then searches for those fingerprints in the new overlay. If it finds
matching segments, those are added to splat config with the appropriate
size. Symbols for those segments are then added to the symbol table.
    
Because this looks at symbols in the context of their segment, symbol
ordering is preserved. These function symbols are added prior to the
`dups` step, which will only add new symbols that have not matches
reference overlays.
    
Currently segments are chosen by finding the largest first, followed by
the most popular. Likewise, symbols names are chosen by popularity.
`OVL_EXPORT` generated symbol names are identified and corrected if they
have a prefix which matches the overlay they were coming from (e.g. if a
symbol starts with `CEN_` in `stcen` it will be updated to match the new
overlay's prefix).
    
Optionally, the `--annotate-symbol-source` option can be passed to
`make-config` to include a comment in the symbol table to track how a
symbol was found (currently only "match" is tracked).