NO3 EntityBreakable data cleanup (#2602)
No changes to the code itself (this one is in so many overlays I'd be
amazed if it wasn't already matching on PSP).

None of the data was pulled in for US yet. But it's identical to the
data for NP3 (which also makes sense).

Found a few ifdef PSP in that data, which could be removed with some
slight tweaks to how the data is written. Also some data splits were
wrong, fixed that too.