NO3 EntityBackgroundBlock improvements (#2605)
I've been making a bit of a personal rule to not do `extern` for data
members on PSP, and it seems to be going well. The way I see it, I
either have to add the data to the symbols file or pull it into the
function, and it's better to have it in the function anyway. So here we
pull in the ObjInit2 data for this function, on both versions.

Also some nice file splitting to prepare things for decompiling the
cutscene information on PSP, which I expect to be a bit of a hefty task.