Re-organise SFX enums (#2608)
I got to share a bit more code between US, HD and PSPEU by leveraging
the ID increment of the sfx enum.

I am also taking this opportunity to explicitly show the ID of the
various entries, to better recognise IDs at glance.

There is a lot I do not yet know. I did not test the BGMs on HD and PSP,
I only worked on this by looking at `src/st/lib_psp/e_shop.c` and
`HandleGameOver`. There are some IDs missing for HD/PSPEU before the
narrator voices. I put a `__ALIGN_ID_341` with a note, waiting for new
IDs to be discovered and for the enum to be completed.