NO3 Cutscene: Decomp on PSP (no codeshare), data/BSS on US (#2610)
This was tricky!

There are enough differences between the two versions that I think it
makes sense to keep these non-codeshared. Therefore I think this will be
the state I'll leave it in, at least for now.

Pulled in data on US though, so at least there's that. The .h file is
looking nicely trimmed down too.