Rework create_entity.h; more PSP NO3 codeshare (#2611)
I'm into a small stretch of PSP NO3 functions that are already
decompiled in very high quality, so very little changes to the C are
necessary.

Big change: I noticed that create_entity.h was almost entirely just a
big sequence of `#include` statements, which were including .h files
that were not included anywhere else. I don't see any purpose to pulling
code into a .h file, if it's only included in a single location. I
believe this is a holdover from when we were originally doing deduping
on a function-by-function basis, and not treating full files as shared
units.

Therefore, I have extracted all the contents of these various .h files,
and placed them into the top-level create_entity.h file sequentially.
Overall, this means the code is easier to follow since we don't have so
many nested levels of .h files.