PSP Maria: match various attack-related entities (#2618)
I hope I did not miss anything. This is the second last PR I made to
decompile the Maria overlay, as there is not much left to match.

I really start to have a full picture of all the various previously
unknown fields and functions. I will probably start documenting this
overlay right with the next coming PR.