Give "opacity" name to Entity.unk6C (#2622)
Turns out this was already discovered in a comment in `ghost.c`. Named
now, both the entity member and the draw flag.

Mostly a find and replace, but not a fully automatic one; I checked as I
went to make sure the member didn't get renamed in any other structs, so
this should just be entities.