Finish decompiling functions for PSP NO3 (#2655)
A bit of a large PR, but most of the changes I made here were small and
inconsequential.

- Very minor trivial backports to make functions match
- Flattening of `st_common.h`
- Pulling in data for US versions of functions, where appropriate
- Pulling out Zombie file for US NO3/NP3 (very surprised this wasn't
done before)
- Codesharing US NO3 files for PSP

NO3 isn't quite done yet - not all data is pulled in, and we need some
cleanup on symbols files and NO3.h (lots of single-file data entries
declared in the .h file), but this PR completes the C code decompilation
for PSP NO3. Very smooth process now that we have so many files that can
just be pulled in as duplicates!