Cleanup former st debug data (#2656)
I went through and added or adjusted the data segments for e_room_bg and
e_red_door. I moved BackgroundBlockInit to e_room_bg and moved
RedDoorTiles to e_red_door, importing the data when necessary. In order
for this to work since g_eRedDoorUV is not in a contiguous location, I
moved it out to it's own file.

The end result of doing all of that allows e_room_bg, e_red_door, and
e_init to be aligned between us and pspeu and merged.