Import `dra` D_801375BC (#2668)
`D_801375BC` takes 4 bytes on PSP indicating the following data is part
of the same array / struct. To fix this I put `D_801375C0` inside the
same struct.
I also made all 'moved' .bss data so far static since those variables
are only used in the same file.
The next variable to be imported is `g_canRevealMap ` according to PSP,
but on PSX that variable is in the `.comm` segment which I have yet to
figure out how to use.