Dockerfile: reduce the number of steps to set-up a container (#2684)
This greatly simplify how developers interact with SOTN using a Docker
container. This essentially skips the need of performing a `make
update-dependencies` the first time a container is created:

* Go is installed inside the image
* Python dependencies and the .venv are installed inside the image

Once the image is built, you can run it by mounting the volumes as usual
and you can get an :ok: by just running `make`. Super simple!

After this PR I plan on adding a CI step to push the image on Docker
Hub, so anyone can pull it to further minimize the development set-up.