Match Remaining PSP Bat Familiar Functions (#2691)
Getting `UpdateBatAttackMode` to work required changing the logic
slightly. The GCC optimized version is interesting in that it will
ignore comparisons that would not produce any code execution. It seems
like it wouldn't be logically equivalent (and if you were relying on
side effects from reads, it wouldn't be), but ends up working out to the
same thing.

`UpdateServantDefault` needed `SetEntityAnimation` to be `static`.