Pulled in code and merged known segments for DAI (#2678)
I have also decompiled e_cutscene, but I'm leaving that for a separate
PR since e_cutscene has a lot going on.

I want to say that I did very little work for this comparatively. This
is 99% just pulling in what has already been done, so thank you to
anybody who decompiled any of the code in these .h files.