Decompiled `dai` `e_cutscene` (#2693)
I did what I could to bring pspeu and us together, but I wasn't able to
get it all the way. I think there is the possibility of merging them
with a couple ifdefs for additional psp code, but it seems to involve
larger scale code adjustments that are beyond my current abilities.

There are two segments associated with the cutscene that seem like they
should be a part of the segment. I tried everything I could think of,
but psx wouldn't accept it. It seems that the cutscene section of code
was likely heavily modified for pspeu, so it is difficult to draw
correlations. This additional segment appears to be code needed to
"tare" the environment for the cutscene, for example changing Alucard to
human form, centering the screen, positioning Alucard, the elevator in
cen, etc.

That being said, I found that psx had a consistent order of:
```
1. CutsceneUnk1 -> SetCutsceneScript
2. CutsceneUnk3 -> CutsceneUnk4
3. DrawCutsceneActorName
4. SetCutsceneEnd -> CutsceneRun -> CutsceneSkip -> ScaleCutsceneAvatar
5. EntityCutScene
6. cutscene_init (three function segment, same order psx and psp)
```
And pspeu had a consistent order of (numbered order psx, visual order
psp):
```
6. cutscene_init (three function segment, same order psx and psp)
0. GetLang
3. DrawCutsceneActorName
2. CutsceneUnk3 -> CutsceneUnk4
1. CutsceneUnk1 -> SetCutsceneScript
4. SetCutsceneEnd -> CutsceneRun -> CutsceneSkip -> ScaleCutsceneAvatar
5. EntityCutScene
```

- I added enum items to PlayerPad for the forms to be used instead of
ifdefs
- I added a small ifdef at the beginning of cutscene_actor_name.h so
that psx and psp can reference the same .h file, even though the code is
different. I intend to add get_lang.h to the psp version in a future PR.
- I created three new .h files consolidating functions using the
consistent subordering
- cutscene_script_text.h, which combines cutscene_unk1.h and
set_cutscene_script.h
- cutscene_script_box.h, which combines cutscene_unk3.h and
cutscene_unk4.h
- cutscene_exec.h (I intend to convert cutscene_run.h to this in a
future PR), which combines set_cutscene_end.h, cutscene_run.h, and
cutscene_scale_avatar.h
- Added cutscene asset extraction for both psx and psp. pspeu required
added the trailing data manually.
- Updated CSOP_SCRIPT_UNKNOWN_4 to be CSOP_HIDE_DIALOG
  - Added CSOP_HIDE_DIALOG  and CSOP_SCRIPT_UNKNOWN_18 macros

I did what made sense to do to begin cleaning up cutscene related code,
but there is still a lot that can be done.