Add new overlay PSP NP3 (#2704)
After finishing NO3, I figured it makes sense for me to move onto NP3
and do the same stuff. This adds the new overlay, and I believe it does
all the standard stuff for a new one, but if I missed something, please
let me know.

I have already opened an Issue to update the steps for a new overlay, so
hopefully this can go more smoothly in the future.