Document map drawing functions (#2718)
There is a master version of the Castle Map, which I'm calling Castle
Blueprint here, which is the version copied from when the Castle Map
revealed by the player is updated. The Castle Blueprint, however, must
be repainted whenever the player changes castles, due to an asymmetry in
what tiles are able to be revealed with normal movement. This only seems
to affect the two Ferryman rooms in Underground Caverns.


![image](https://github.com/user-attachments/assets/7cd93f14-e778-445e-92b4-cb3dca14d467)

Credit to @Xeeynamo for help in finding the map-drawing functions I was
looking for.