Document Jewel Sword drops (#2730)
Note that the item IDs used in the table for MAD are different. They
appear to be off by 4 entries. I don't know of a way to verify the Jewel
Sword behavior in MAD since there are no enemies in the stage, but I
believe the new item IDs correspond to Lapis Lazuli, Ring of Ares, Gold
Ring, and Silver Ring.

I also removed an unused variable in `HitDetection` that I noticed while
renaming `randCompare`.