Add `TOP` Overlay to `us` (#2741)
Adds `TOP` and dedups shared segments.

`TOP` is a pretty simple stage with only a few unique entities and a
cutscene.

`TOP`'s `CutsceneRun` implementation was the same as `CEN`'s. I've
updated `cutscene_run.h` to support both.

`EntityBlueAxeKnight` is identical to `NO1`'s. I've moved `NO1`'s
implementation to a shared header which is now used for both stages.

`EntityBackgroundBlock` was the same version used in `NO4`.