Shared `e_room_bg` (#2790)
Most stages use one of a few variants of `e_room_bg`. Those differences
can be enabled via macros depending on the implementation.

The first variation is how `facingLeft` is assigned. Many overlays
assign that to `unk5A`, but others assign it to `facingLeft` and then
have an additional copy of `unk5A`. There's a possibility that future
PSP versions will also need `BG_FACING_LEFT_FIX` along with an unsigned
load, but the current overlays do not.

The other variation is whether or not `scaleX` is set. So far, any
overlay that sets the scale uses the same value.