Relicense psy-q as MIT (#2797)
Interest in psy-q decompilation has increased a lot recently so I wanted
to get the licensing situation cleaned up for the SDK part of the repo.
The contributors are:

```
git shortlog -n -s -- src/main/psxsdk
    79  sozud
    20  Luciano Ciccariello
     8  gamezter
     5  Jonathan Hohle
     4  Ross Squires
     4  ser-pounce
     3  Tuomas Tynkkynen
     2  Josh Schreuder
     1  Alejandro Asenjo
     1  Alejandro Asenjo Nitti
     1  Formatting bot
     1  Josh Lory
     1  Mark Street
     1  Onenutmcgee
```

My rationale for MIT is:
The license is approved both by the FSF and OSC:
https://spdx.org/licenses/
https://directory.fsf.org/wiki/License:Expat
https://opensource.org/license/mit

It makes sense to have a permissive license so this work can be used in
any repo and also as the basis for projects like psyz
https://github.com/psyqz/psyz/tree/main/decomp/src

FSF recommends Apache 2.0 for permissive works but I decided against it
since it's not GPLv2 compatible.

Any issues with this? Need approval from each of the following:
@Xeeynamo 
@gamezter 
@hohle 
@rossimo 
@ser-pounce 
@dezgeg 
@JoshSchreuder 
@sonicdcer 
@joshlory 
@mkst 
@Onenutmcgee