Fix Duplicates Report, Analyze Calls, & Function Finder (#2824)
The duplicates report is no longer picking up dynamic symbols, which
represent an increasing percentage of all function symbols. Previously
generating the report would also overwrite files in the work tree which
made it difficult to run ad-hoc. `make duplicates-report` should work in
any working tree without modifying any files outside of `asm` or
`build`.

`symbols.py` now has a `dynamic` command that will dump all of the
symbols in an elf file that are not defined in symbol tables for that
overlay. This uses the same filtering rules as the `elf` command. Any
symbols with the same name or address of one of the static symbols will
be excluded from the output.

The `force_symbols` make target emits a table of dynamic symbols for
each of the `us` overlays and stores them in `$BUILD_DIR/config`. This
"dynamic" symbol file is appended to the static symbol file list of each
overlay at extract time. This makes it possible to build the duplicates
report without moving `src` or overwriting anything in `config`.

`analyze_calls.py` has been updated with current rules based on new
patterns and supports mapping matching functions.

Reports are now written to the build directory to keep the working
directory clean. The github workflow moves them to the expected place
for the `gh-duplicates` branch.