Build All Versions Together (#2873)
Changes the default target of `make` and `tools/builds/gen.py` to build
all decomped versions of the game together in one bulid graph.

There are two parts to this change: making assets writes atomic to
allowing assets to be shared across versions, and updating the build
graph to not duplicate shared targets.

Different versions of the game are currently configured to write to the
same asset paths. I think this is a good thing™ since it makes differing
assets obvious. The work to separate these assets was done in
4bd3e99dbca32d5548ebf08e48b70abbbc9de168 but running multiple versions
of `sotn-assets` in parallel resulted in race conditions that prevented
multiple versions of the game from building together (whether in the
same graph or not). This change updates `sotn-assets` to write files
atomically which avoid partial reads while allowing multiple instances
of `sotn-assets` to run concurrently. An alternative to this approach
would be to put `sotn-assets` rules in the same ninja pool, but that
would force them to be serialized, which seems unnecessary, but possibly
simpler.

The second area is merging all three build graphs into a single
monolithic graph. This is primarily meant eliminating duplicate asset
steps for shared assets.

`make` will now build `us`, `hd`, and `pspeu`. using the `VERSION` env
var still works and will only build that version. `make clean` cleans
all versions.

picci builds are out of scope for this change.