Support Building All Versions Together (#2946)
Restores part of fba6cac4531afffca5228762d04efde2a6e506f9 (#2873)

There are three parts to this change: making assets writes atomic to
allowing assets to be shared across versions, supporting subdirectories
for palettes and spritebanks, and updating the build graph to not
duplicate shared targets.

Different versions of the game are currently configured to write to the
same asset paths. I think this is a good thing™ since it makes differing
assets obvious. The work to separate these assets was done in
4bd3e99dbca32d5548ebf08e48b70abbbc9de168 but running multiple versions
of `sotn-assets` in parallel resulted in race conditions that prevented
multiple versions of the game from building together (whether in the
same graph or not). This change updates `sotn-assets` to write files
atomically which avoid partial reads while allowing multiple instances
of `sotn-assets` to run concurrently.

Between the earlier assets work and now, DAI has added palettes and
spritebanks that differ between US and PSP. Those asset handlers have
been updated to support writing to subdirectories.

The third part is merging all three build graphs into a single
monolithic graph. This is primarily meant eliminating duplicate asset
steps for shared assets.

Unlike the previous version of this change, the default target has not
been changed and this will not run in CI yet. `build_all` and
`clean_all` targets have been added that can be invoked explicitly.