Add objdiff support (#2988)
I am expanding `sotn-assets` to orchestrate more complex parts of our
build-chain and tooling. This PR is specifically for the objdiff
support, which will be crucial later on to start reporting progress to
decomp.dev . Note that only one version at the time is allowed.

### `VERSION=us go run ./tools/sotn-assets objdiff-gen`

Generates a `objdiff.json` to the repo's root for objdiff to operate to.

### `VERSION=us go run ./tools/sotn-assets objdiff-gui`

Generates a `objdiff.json` and opens the ObjDiff GUI.

### `VERSION=us go run ./tools/sotn-assets objdiff dra/4A538
func_800EAD0C`

Generates a `objdiff.json` and invoke the interactive version of ObjDiff
CLI. One potential improvement here is to infer the `dra/4A538` part,
and provide a simple TUI to select the right overlay in case of a symbol
collision.

### Shorter version

You might want to create an `alias` to invoke the tool more quikcly,
like `alias sotdiff="go run ./tools/sotn-assets objdiff"` or similar, so
you can do `sotdiff dra/4A538 func_800EAD0C`.