Link main dra symbols (#3051)
On PSP, I noticed a discrepancy between the symbols we have in
`main.exe` and those used by the overlays. This disconnection caused
functions and extern data to be mis-documented or to have mis-matching
declaration. It was also not possible to rename an exported main symbol
that is used by other overlays.

This PR solves all the problems described above by creating a linker
script from `main.elf` that gets used by all other overlays when
linking. The downside is if `main.elf` does not get produced, no other
overlays could be compiled. A `main.elf` that produces shifted symbols
will result to all other overlays to not match. This is by design.

Removing the `undefined_syms_auto` and `undefined_funcs_auto` for DRA
enforces the linker for symbols to be correctly synchronized. What this
means, before the symbol `D_880000` in DRA was defined by the Splat
generated `undefined_syms_auto`, which was helpful when generating
DRA.ELF. But without the use of `undefined_syms_auto`, the symbol must
come from our generated `build/pspeu/main.symexport.ld` which has the
correct name D_psp_0880000`. Again, this is done to ensure correctness.

It's as big change, but the meat of these changes is in
`tools/builds/gen.py`.