fname: decompile a good chunk of code (#3056)
It's a TT_000 beta, there's a lot of copy&paste from that overlay. It is
linked to a pre-release of `main.exe` which means most shared symbols
need to be rebased on the new `config/symbols.beta_fname.txt`.

I will attempt to merge the two overlays once this `fname` is fully
matching.

The development timeline, from older to newest, seems to be:

SD overlay > MAD stage > FNAME beta familiar > HD beta build > JP 1.0